+++ .release-it.ts
@@ -1,6 +1,5 @@
 import type {Config} from 'release-it';
 
-// eslint-disable-next-line import/no-default-export
 export default {
   git: {
     requireCleanWorkingDir: false,